Move Floyd-specific god-mode wandering toggle out of the engine (PR feedback)
GodModeProcessor lives in GameEngine and must not know about Floyd, who is
Planetfall-specific. Replace the Floyd-named interface with a generic
IGodModeCommandHandler extension point: the engine's god-mode processor
delegates unrecognized commands to the game context, and PlanetfallContext
now owns the "god mode [no] wander" parsing, the Floyd wording, and the
FloydWanderingDisabled flag. FloydMovementManager checks PlanetfallContext
directly, the same pattern Floyd.Act() already uses for its per-turn flags.

Player-visible behavior is unchanged; all existing tests pass as-is.
