+++ Planetfall/Item/Kalamontee/Mech/FloydPart/FloydMovementManager.cs
@@ -1,5 +1,4 @@
 using Model.AIGeneration;
-using Model.Interface;
 using Planetfall.Location;
 
 namespace Planetfall.Item.Kalamontee.Mech.FloydPart;
@@ -43,7 +42,7 @@ public class FloydMovementManager(Floyd floyd)
         {
             // Random chance to no longer follow (1 in 5 chance) - skipped when god mode has disabled
             // Floyd's random wandering for deterministic playtesting.
-            var wanderingDisabled = context is IFloydWanderingContext { FloydWanderingDisabled: true };
+            var wanderingDisabled = context is PlanetfallContext { FloydWanderingDisabled: true };
             if (!wanderingDisabled && floyd.Chooser.RollDiceSuccess(5))
             {
                 floyd.IsOffWandering = true;
@@ -93,7 +92,7 @@ public class FloydMovementManager(Floyd floyd)
             return null;
 
         // God mode may have disabled Floyd's random wandering for deterministic playtesting.
-        if (context is IFloydWanderingContext { FloydWanderingDisabled: true }) return null;
+        if (context is PlanetfallContext { FloydWanderingDisabled: true }) return null;
 
         if (!floyd.Chooser.RollDiceSuccess(20)) return null;
